Changed from size_t to unsigned long int

+++ handle_pointers.c
@@ -15,7 +15,8 @@ int handle_pointer(__attribute__((unused)) const format_specifier * spec,
 	char hex_str[17];
 	int characters_printed;
 	size_t initial_length = buffer->length;
-	size_t pointer = (size_t)va_arg(args, size_t *);
+	unsigned long int pointer = (unsigned long int) va_arg(
+		args, unsigned long int *);
 
 	utob(pointer, hex_str, HEX);
 	append_string(buffer, "0x");
